Move tree components under the tree folder/package
